Auto install corrections

+++ autormcp.lua
@@ -17,8 +17,11 @@ local folders = {nvim='~/.config/nvim/',
 -- Files
 for key, value in pairs(files) do
     if arg[1] == 'install' then
-        os.execute("rm -f" .. value .. key)
-        os.execute("cp " .. PATH_FILES .. key .. " " .. value .. key)
+        if value ~= "~/" then
+            os.execute("rm -f" .. value .. key)
+        end
+        os.execute("mkdir " .. value)
+        os.execute("cp -r " .. PATH_FILES .. key .. " " .. value .. key)
     else
         os.execute("rm " .. PATH_FILES .. key)
         os.execute("cp " .. value .. key .. " " .. PATH_FILES .. key)
@@ -30,6 +33,7 @@ end
 for key, value in pairs(folders) do
     if arg[1] == 'install' then
         os.execute("rm -rf " .. value .. key)
+        os.execute("mkdir " .. value)
         os.execute("cp -r " .. CURRENT_PATH .. key .. "/* " .. value)
     else
         os.execute("rm -rf " .. CURRENT_PATH .. key)
